pyproject.toml
[project]
name = "lx-annotate"

version = "0.0.1"
description = "This webapp provides a frontend interface for annotating AI generated output."
authors = [
    { name = "Max Hild", email = "maxhild10@gmail.com"},
    { name = "Thomas J. Lux", email = "lux_t1@ukw.de" },
    { name = "Hamza Zahid" },
]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3.12",
  "Topic :: Scientific/Engineering :: Medical Science Apps.",
]
maintainers = [
    { name = "Max Hild", email = "maxhild10@gmail.com"},
    { name = "Thomas J. Lux", email = "lux_t1@ukw.de" },
    { name = "Hamza Zahid" },
    { name = "Peter Kowalczyk" },
]
license = "MIT"
readme = "README.md"
requires-python = ">=3.12.0,<3.13"

dependencies = [
    "django>=5.1.1",
    "gunicorn>=23.0.0",
    "djangorestframework>=3.12.4",
    "django-cors-headers>=4.0.0",
    "django-webpack-loader>=0.7.0",
    "requests>=2.32.3",
    "django-vite>=3.0.6",
    "whitenoise>=6.8.2",
    "django-extensions>=3.2.3",
    "python-dotenv>=1.0.1",
    "python-environ>=0.4.54",
    "pytest>=8.3.5",
    "pytest-django>=4.11.1",
    "django-environ>=0.12.0",
    "django-rest-framework>=0.1.0",
    "watchdog>=3.0.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "transformers>=4.44.0,<4.45.0",
    "pillow>=9.0.0",
    "numpy>=1.21.0",
    "celery>=5.5.3",
    "drf-spectacular>=0.28.0",
    "django-modeltranslation>=0.19.16",
    "daphne>=4.2.1",
    "psycopg[binary,pool]>=3.2.12",
    "mozilla-django-oidc>=4.0.1",
    "pre-commit>=4.5.0",
    "de-core-news-lg",
    "django-stubs>=5.2.8",
    "types-requests>=2.32.4.20250913",
    "black>=25.12.0",
    "mypy>=1.18.2",
    "endoreg-db==0.8.9.16",
    "pydantic-settings>=2.12.0",
    "pydantic>=2.12.5",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "lx_annotate.settings"
python_files = ["test_*.py", "*_test.py", "tests/python/*.py"]
addopts = [
  "--import-mode=importlib",
  "-ra",
]

[tool.uv.sources]
de-core-news-lg = { url = "https://github.com/explosion/spacy-models/releases/download/de_core_news_lg-3.8.0/de_core_news_lg-3.8.0-py3-none-any.whl" }

[project.optional-dependencies]
dev = [
  "build>=1.3.0",
  "flake8>=7.3.0",
  "linkify-it-py>=2.0.3",
  "myst-parser>=4.0.1",
  "mypy>=1.16.0",
  "pip>=25.2",
  "pre-commit>=4.5.0",
  "pytest>=9.0.2",
  "pytest-cov>=6.1.1",
  "pytest-env>=1.1.5",
  "pytest-mypy-plugins>=3.2.0",
  "pytest-xdist>=3.8.0",
  "sphinx-autodoc-typehints>=3.5.2",
  "sphinx-copybutton>=0.5.2",
  "sphinx>=8.2.3,<9.0",
  "furo>=2025.9.25",
  "ruff>=0.14.8",
  "twine>=6.2.0",
]

docs = [
  "linkify-it-py>=2.0.3",
  "myst-parser>=4.0.1",
  "sphinx-autodoc-typehints>=3.5.2",
  "sphinx-copybutton>=0.5.2",
  "sphinx>=8.2.3,<9.0",
  "furo>=2025.9.25",
]

[tool.ruff]
line-length = 88

[tool.black]
line-length = 88
target-version = ["py312"]
[tool.mypy]
ignore_missing_imports = true
exclude = ["venv", ".devenv"]

[[tool.mypy.overrides]]
module = [
    "rest_framework.*",
    "environ.*",
    "whitenoise.*",
    "mozilla_django_oidc.*"
]
ignore_missing_imports = true

# [tool.uv.sources]
# endoreg-db = { workspace = true }
# lx-anonymizer = { workspace = true }

# [tool.uv.workspace]
# members = [
#    "libs/endoreg-db",
#    "libs/lx-anonymizer",
# ]
